pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-society-sim"
version = "0.1.0"
description = "A free-form multi-agent simulation of AI-driven social transition."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.0"
]

[project.scripts]
ai-society-sim = "ai_society_sim.cli:main"

[tool.setuptools.packages.find]
where = ["src"]